+++ nix/packages/nxr.nix
@@ -8,7 +8,7 @@
 
 rustPlatform.buildRustPackage {
   pname = "nxr";
-  version = "2.3.1";
+  version = "2.3.2";
 
   inherit src;